requirements.txt
requests
python-dotenv
pandas
rich
flask
pyyaml
